adds disabled button
