+++ app/src/components/ScheduleAvailableDays/Day.js
@@ -3,15 +3,16 @@ import classes from './Day.module.scss';
 
 export default function Day({ day, selected, dispatch, idx }) {
   return (
-    <div className={classes.day}>
-    <input
-      type='checkbox'
-      checked={selected}
-      onChange={(e) =>
-        dispatch({ type: 'TOGGLE_SELECTED', payload: idx })
-      }
-    />
-    {day}
-  </div>
+    <div
+      className={classes.day}
+      onClick={(e) => dispatch({ type: 'TOGGLE_SELECTED', payload: idx })}
+    >
+      <input
+        type='checkbox'
+        checked={selected}
+        onChange={(e) => dispatch({ type: 'TOGGLE_SELECTED', payload: idx })}
+      />
+      {day}
+    </div>
   );
 }
