+++ app/src/components/ScheduleAvailableDays/ScheduleAvailableDays.module.scss
@@ -9,3 +9,8 @@
     flex-direction: row;
 }
 
+.disabled {
+    background: gray;
+    cursor: wait;
+}
+
