+++ app/src/components/ScheduleAvailableDays/WeekDays.js
@@ -1,22 +1,11 @@
-import React, { useReducer } from 'react';
+import React from 'react';
 import Day from './Day';
-import { weekDays } from '../../store/index';
-import { reducer } from './reducer';
-const initialState = { weekDays };
 
-export default function WeekDays() {
-  const [state, dispatch] = useReducer(reducer, initialState);
+export default function WeekDays({ dispatch, state }) {
   return (
     <>
-      <div className='body-text'>
-        <p>What days are you typically available?</p>
-        <p>Select from below.</p>
-        <p onClick={() => dispatch({ type: 'TOGGLE_ALL' })}>
-          Select / Unselect All
-        </p>
-      </div>
       {state.weekDays.map(({ day, selected }, idx) => (
-        <Day day={day} selected={selected} idx={idx} dispatch={dispatch} />
+        <Day day={day} selected={selected} idx={idx} key={idx} dispatch={dispatch} />
       ))}
     </>
   );
